+++ lua/campaign_main.lua
@@ -27,6 +27,7 @@ wc2x.ai_director.init(wc2x.config)
 wc2x.debug = wc2x.ai_director.debug
 wc2x.debug_panel = wesnoth.dofile("./wc2x/debug_panel.lua")
 wc2x.debug_panel.init(wc2x.config)
+wc2x.boss_roster = wesnoth.dofile("./wc2x/boss_roster.lua")
 wc2x.final_boss = wesnoth.dofile("./wc2x/final_boss.lua")
 wc2x.final_boss.init(wc2x.config)
 wc2x.unit_finder = wesnoth.dofile("./wc2x/unit_finder.lua")